Project : Monster Rolodex
Reviewing the React Basics by creating a monster rolodex appication

Change the generated App function-based component to a class-based component
Create a state monsters and initialize it to an empty array
define a componentDidMount() life-cycle method in which we fetch a list of person
in the render method that renders the App component
we insert a CardList component that we import in the App.js file
The CardList component has the monster attribute with the value of the state.monster object created

Create the CardList function-based component that takes in props as parameter
CardList is created in a card-list directory which is itself inside the components directory created under the src
The CardList component renders a div. Inside that div we use a javascript expression and we iterate over the monsters array recover from the props parameter. What we iterate is a Card component that we create inside the card component in the components directory.